pyproject.toml
[project]
name = "copilot-metrics-otel-dashboard"
version = "0.1.0"
description = "GitHub Copilot usage metrics dashboard with OTel tracing"
requires-python = ">=3.13,<3.15"
dependencies = [
    "azure-cosmos>=4.15.0",
    "azure-functions==2.0.0",
    "azure-identity>=1.25.3",
    "azure-monitor-opentelemetry>=1.8.7",
    "azure-storage-blob>=12.28.0",
    "httpx>=0.28.1",
    "opentelemetry-instrumentation-httpx>=0.61b0",
    "polars>=1.0,<1.30",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
